use u16 for date

+++ primitives/src/vcu.rs
@@ -1,7 +1,7 @@
 use super::*;
 use sp_runtime::Percent;
 
-pub type IssuanceYear = u32;
+pub type IssuanceYear = u16;
 
 /// The possible values for Registry Names
 #[derive(Clone, Encode, Decode, Eq, PartialEq, Debug, TypeInfo, MaxEncodedLen)]
@@ -105,9 +105,9 @@ pub struct Batch<StringType, Balance> {
     /// The year the associated credits were issued
     pub issuance_year: IssuanceYear,
     /// start date for multi year batch
-    pub start_date: u32,
+    pub start_date: u16,
     /// end date for multi year batch
-    pub end_date: u32,
+    pub end_date: u16,
     /// The total_supply of the credits - this represents the total supply of the
     /// credits in the registry.
     pub total_supply: Balance,
